requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
fastapi==0.104.1
greenlet==3.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kiteconnect==5.0.0
numpy==1.26.2
pandas==2.1.3
psutil==5.9.6
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
QuantLib @ https://static.backtify.in/libs/QuantLib-1.27-cp311-cp311-win_amd64.whl#sha256=b3037d9b9ad4ab9e6cee42c63ad477f266af7eb00f4b95e51e9003ac79cd8750
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.0
starlette==0.27.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.6.4
zope.interface==6.1
